feat: domain associateHouseToUser

+++ src/main/kotlin/com/example/modules/locals/presenter/dto/house/request/AssociateHouseToUserRequestDto.kt
@@ -1,6 +1,6 @@
 package com.example.modules.locals.presenter.dto.house.request
 
 data class AssociateHouseToUserRequestDto (
-    val houseId: Int,
-    val residentId: Int
+    val houseId: Long,
+    val residentId: Long
 )
